+++ src/theme/ThemeProvider.tsx
@@ -1,12 +1,12 @@
-import { defaultTheme } from "@isonia/theme-default";
 import type { PropsWithChildren } from "react";
 import { useEffect } from "react";
+import { appCoreTheme } from "./isonia-theme";
 
 export function ThemeProvider({ children }: PropsWithChildren): JSX.Element {
   useEffect(() => {
     const root = document.documentElement;
-    root.dataset.isoniaTheme = defaultTheme.id;
-    for (const [name, value] of Object.entries(defaultTheme.cssVariables)) {
+    root.dataset.isoniaTheme = appCoreTheme.id;
+    for (const [name, value] of Object.entries(appCoreTheme.cssVariables)) {
       root.style.setProperty(name, value);
     }
   }, []);